requirements.txt
streamlit
python-dotenv
google-generativeai
sklearn
nltk
numpy